+++ src/ldc/supervised/classification/__init__.py
@@ -1,0 +1,3 @@
+from ._core import ClassificationData, ClassificationFilter, ClassificationReader, BatchClassificationWriter, StreamClassificationWriter
+from ._jsonlines import JsonLinesClassificationReader, JsonLinesClassificationWriter
+from ._classification_label_map import ClassificationLabelMap