+++ CHANGES.rst
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+0.1.1 (????-??-??)
+------------------
+
+- added `classification` domain
+- added `from-jsonlines-cl` reader and `to-jsonlines-cl` writer for classification data in JSON lines format
+- added filter `pretrain-sentences-to-classification` to turn pretrain data into classification data (with a predefined label)
+- added filter `classification-label-map` that can generate a label string/int map
+
+
 0.1.0 (2024-02-05)
 ------------------